requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.5.0  # https://github.com/evansd/whitenoise
redis==4.6.0  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.1  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==2.0.0  # https://github.com/mher/flower
uvicorn[standard]==0.22.0  # https://github.com/encode/uvicorn
rich==13.5.0
markdown==3.4.4
boto3==1.28
rich==13.5
pydantic<2.0  # Less than 2.0 because of django pydantic field
django-pydantic-field==0.2.11
sentry-sdk==1.40.4  # https://github.com/getsentry/sentry-python
django-cachalot

# Django
# ------------------------------------------------------------------------------
django==4.2.10  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
# django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.3.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.1.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.26.3  # https://github.com/tfranzel/drf-spectacular
django-filter==23.2  # https://github.com/carltongibson/django-filter
django-storages[boto3]==1.14.2  # https://github.com/jschneier/django-storages
# Authentication API
djoser==2.2.0
# Email sending
django-sendgrid-v5==1.2.2
django-anymail[sendgrid]==10.0  # https://github.com/anymail/django-anymail

## Formerly dev-only dependencies
# However we cannot run the app without some of these these dependencies
# and it is easier to maintain in a single file
# @TODO move to Python Poetry or pyproject.toml for dependencies
# ------------------------------------------------------------------------------

Werkzeug[watchdog]==2.3.6 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb
psycopg[binary]==3.1.9  # https://github.com/psycopg/psycopg
watchfiles==0.19.0  # https://github.com/samuelcolvin/watchfiles

# Testing
# ------------------------------------------------------------------------------
mypy==1.4.1  # https://github.com/python/mypy
django-stubs==4.2.2  # https://github.com/typeddjango/django-stubs
boto3-stubs[s3,core]
mypy_boto3_s3==1.28.19
pytest==7.4.0  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.7  # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs==3.14.2  # https://github.com/typeddjango/djangorestframework-stubs

# Documentation
# ------------------------------------------------------------------------------
sphinx==6.2.1  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2021.3.14 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==6.0.0  # https://github.com/PyCQA/flake8
flake8-isort==6.0.0  # https://github.com/gforcada/flake8-isort
coverage==7.2.7  # https://github.com/nedbat/coveragepy
black==23.3.0  # https://github.com/psf/black
djlint==1.31.1  # https://github.com/Riverside-Healthcare/djLint
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==3.3.3  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.1  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==4.1.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.0.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django


## Formerly production-only dependencies
# @TODO move to Python Poetry or pyproject.toml for dependencies
# ------------------------------------------------------------------------------

newrelic==9.6.0
gunicorn==20.1.0  # https://github.com/benoitc/gunicorn
# psycopg[c]==3.1.9  # https://github.com/psycopg/psycopg
